Data used in student exercises for RDSS based on LAPOP survey of Brazil in 2018

These data were resampled with replacement from LAPOP data (to 10,000 rows) for a subset of variables. These data cannot be used for scientific inferences, and are only useful for teaching purposes. ID numbers were scrambled so that individuals and municipalities cannot easily be identified.

Usage

lapop_brazil

Format

A data.frame

Details

Download the original data from https://www.vanderbilt.edu/lapop/raw-data.php

See https://www.vanderbilt.edu/lapop/core-surveys.php for survey questionnaire